Change to compiler to enforce relative pathing
When plugging in an sdcard and installing a script on eMMC the compiler would break due to a relative/absolute pathing structure.

I think it tries to look in something like STORAGETYPE:/scripts/rfsuite.

Forcing the compiler to use no absolute paths solves the issue.

I dont like it and want a proper fix!