Optimised modelimage, image and utils.lua
These objects now use the same utils.box as other objects, keeping this in one area within utils when it comes to drawing. Optimised util.lua so it now has fudgefactor to allow better placement of text that contain symbols and ensures displayvalue is also centered dynamically based on the titlefont size.