Fuel logic with reserve %
Update fuel telemetry sensor so its now a 'virtual' sensor and has reserve % functionality. Adjusted sim sensor for fuel to be use the same logic as the main sensor, modified audio alert for Charge Level to trigger when the value is 10% or below and not based on consmptionwarningpercentage anymore as this is the new value for setting a Reserve %. Will raise another PR for renaming consumptionwarningpercentage so it reflects the fact it is now actually reserve %. Tested everything in the sim, alerts trigger as expected when below 10% fuel with charge level audio alert turned on and fuel level now factors in the new reserve % defined under battery settings.